fix some signed/unsigned mismatch (#828)
* fix unsigned int in select_battle_command

* fix duel::set_responsei should use signed int

* fix duel::write_buffer should use size_t

* fix Debug.SetAIName & ShowHint truncation logic

* fix SIZE_* constant

+++ duel.cpp
@@ -121,7 +121,7 @@ void duel::restore_assumes() {
 		pcard->assume_type = 0;
 	assumes.clear();
 }
-void duel::write_buffer(const void* data, int size) {
+void duel::write_buffer(const void* data, size_t size) {
 	vector_write_block(message_buffer, data, size);
 }
 void duel::write_buffer32(uint32_t value) {
@@ -136,7 +136,7 @@ void duel::write_buffer8(uint8_t value) {
 void duel::clear_buffer() {
 	message_buffer.clear();
 }
-void duel::set_responsei(uint32_t resp) {
+void duel::set_responsei(int32_t resp) {
 	game_field->returns.ivalue[0] = resp;
 }
 void duel::set_responseb(byte* resp) {
